requirements.txt
Django==1.7.1
django-debug-toolbar==1.2.2
django-filter==0.9.1
django-model-utils==2.2
#django-notifications-hq==0.6.2
-e git+https://github.com/rizotas/django-notifications.git#egg=notificationsPython3
django-user-sessions==1.1.0
djangorestframework==3.0.2
sqlparse==0.1.14
